azure-function-g4f/requirements.txt
g4f==0.3.8.4
pymongo==4.6.1
azure-functions
